Update README.md
As a part of the guided project from Youtube, this project is aimed to create charts for the analysis of sales data. Pivot tables were created based on a number of parameters and charts were built to visualize the same. The charts were combined to build the dashboard.

Based on the dataset, pivot tables and charts were created to analyse the following:

Sales vs orders, Sales: Top 5 states, Order status, Sales: Men vs Women, Orders: Age vs Gender, Orders: Channels